+++ MiscHelpers/Common/PanelView.cpp
@@ -112,7 +112,7 @@ QStringList CPanelView::CopyHeader()
 	QAbstractItemModel* pModel = GetModel();
 	QTreeView * pView = GetView();
 
-	QStringList Headder;
+	QStringList Header;
 	for (int i = 0; i < pModel->columnCount(); i++)
 	{
 		if (/*!m_CopyAll &&*/ pView->isColumnHidden(i) && !m_ForcedColumns.contains(i))
@@ -120,9 +120,9 @@ QStringList CPanelView::CopyHeader()
 		QString Cell = pModel->headerData(i, Qt::Horizontal, Qt::DisplayRole).toString();
 		if (!m_SimpleFormat)
 			Cell = "|" + Cell + "|";
-		Headder.append(Cell);
+		Header.append(Cell);
 	}
-	return Headder;
+	return Header;
 }
 
 QStringList CPanelView::CopyRow(const QModelIndex& ModelIndex, int Level)
@@ -176,11 +176,11 @@ void CPanelView::OnCopyPanel()
 	FormatAndCopy(DumpPanel());
 }
 
-void CPanelView::FormatAndCopy(QList<QStringList> Rows, bool Headder)
+void CPanelView::FormatAndCopy(QList<QStringList> Rows, bool Header)
 {
 	int RowCount = Rows.length();
 
-	if (Headder)
+	if (Header)
 	{
 		Rows.prepend(QStringList());
 		Rows.prepend(CopyHeader());
@@ -188,19 +188,19 @@ void CPanelView::FormatAndCopy(QList<QStringList> Rows, bool Headder)
 	}
 
 	QStringList TextRows;
-	if (m_SimpleFormat || !Headder)
+	if (m_SimpleFormat || !Header)
 	{
 		foreach(const QStringList& Row, Rows)
 			TextRows.append(Row.join(m_CellSeparator));
 	}
-	else if(Rows.size() > (Headder ? 3 : 0))
+	else if(Rows.size() > (Header ? 3 : 0))
 	{
-		int Columns = Rows[Headder ? 3 : 0].count();
+		int Columns = Rows[Header ? 3 : 0].count();
 		QVector<int> ColumnWidths(Columns, 0);
 
 		foreach(const QStringList& Row, Rows)
 		{
-			for (int i = 0; i < Min(Row.count(), Columns); i++)
+			for (int i = 0; i < std::min(Row.count(), Columns); i++)
 			{
 				int CellWidth = Row[i].length();
 				if (ColumnWidths[i] < CellWidth)
@@ -217,12 +217,12 @@ void CPanelView::FormatAndCopy(QList<QStringList> Rows, bool Headder)
 					bool More = false;
 
 					QString RowText;
-					for (int i = 0; i < Min(Row.count(), Columns); i++) 
+					for (int i = 0; i < std::min(Row.count(), Columns); i++) 
 					{
 						if (Row[i].length() > Pos)
-							RowText.append(Row[i].mid(Pos, m_MaxCellWidth).leftJustified(Min(m_MaxCellWidth, ColumnWidths[i]) + 3));
+							RowText.append(Row[i].mid(Pos, m_MaxCellWidth).leftJustified(std::min(m_MaxCellWidth, ColumnWidths[i]) + 3));
 						else
-							RowText.append(QString(Min(m_MaxCellWidth, ColumnWidths[i]) + 3, ' '));
+							RowText.append(QString(std::min(m_MaxCellWidth, ColumnWidths[i]) + 3, ' '));
 
 						if (Row[i].length() > Pos + m_MaxCellWidth)
 							More = true;
@@ -236,7 +236,7 @@ void CPanelView::FormatAndCopy(QList<QStringList> Rows, bool Headder)
 			else
 			{
 				QString RowText;
-				for (int i = 0; i < Min(Row.count(), Columns); i++)
+				for (int i = 0; i < std::min(Row.count(), Columns); i++)
 					RowText.append(Row[i].leftJustified(ColumnWidths[i] + 3));
 				TextRows.append(RowText);
 			}